delete brick_value.xml, write value-Strings in strings.xml, short changes in some bricks

+++ catroid/src/org/catrobat/catroid/common/BrickValues.java
@@ -1,24 +1,24 @@
 /**
- *  Catroid: An on-device visual programming system for Android devices
- *  Copyright (C) 2010-2013 The Catrobat Team
- *  (<http://developer.catrobat.org/credits>)
- *  
- *  This program is free software: you can redistribute it and/or modify
- *  it under the terms of the GNU Affero General Public License as
- *  published by the Free Software Foundation, either version 3 of the
- *  License, or (at your option) any later version.
- *  
- *  An additional term exception under section 7 of the GNU Affero
- *  General Public License, version 3, is available at
- *  http://developer.catrobat.org/license_additional_term
- *  
- *  This program is distributed in the hope that it will be useful,
- *  but WITHOUT ANY WARRANTY; without even the implied warranty of
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- *  GNU Affero General Public License for more details.
- *  
- *  You should have received a copy of the GNU Affero General Public License
- *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ * Catroid: An on-device visual programming system for Android devices
+ * Copyright (C) 2010-2013 The Catrobat Team
+ * (<http://developer.catrobat.org/credits>)
+ * 
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ * 
+ * An additional term exception under section 7 of the GNU Affero
+ * General Public License, version 3, is available at
+ * http://developer.catrobat.org/license_additional_term
+ * 
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU Affero General Public License for more details.
+ * 
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 package org.catrobat.catroid.common;
 
@@ -32,24 +32,24 @@ public class BrickValues {
 	public static final int Y_POSITION = 200;
 	public static final int CHANGE_X_BY = 10;
 	public static final int CHANGE_Y_BY = 10;
-	public static final Float Move_Steps_Value = (float) 10;
-	public static final Float TURN_REIGTH = (float) 15;
-	public static final Float POINT_IN_DIRECTION = (float) 90;
+	public static final Float Move_Steps_Value = 10f;
+	public static final Float TURN_REIGTH = 15f;
+	public static final Float POINT_IN_DIRECTION = 90f;
 	public static final int GLIDE_SECONDS = 1000;
 	public static final int GO_BACK = 1;
 	public static final String POINT_TOWARDS = "Nothing...";
 
 	//constants Looks
-	public static final Float SET_SIZE_TO = (float) 60;
-	public static final Float CHANGE_SIZE_BY = (float) 10;
-	public static final Float SET_GHOST_EFFECT = (float) 50;
-	public static final Float CHANGE_GHOST_EFFECT = (float) 25;
-	public static final Float SET_BRIGHTNESS_TO = (float) 50;
-	public static final Float CHANGE_BRITHNESS_BY = (float) 25;
+	public static final Float SET_SIZE_TO = 60f;
+	public static final Float CHANGE_SIZE_BY = 10f;
+	public static final Float SET_GHOST_EFFECT = 50f;
+	public static final Float CHANGE_GHOST_EFFECT = 25f;
+	public static final Float SET_BRIGHTNESS_TO = 50f;
+	public static final Float CHANGE_BRITHNESS_BY = 25f;
 
 	//constants Sounds
-	public static final Float SET_VOLUMEN_TO = (float) 60;
-	public static final Float CHANGE_VOLUMEN_BY = (float) -10;
+	public static final Float SET_VOLUMEN_TO = 60f;
+	public static final Float CHANGE_VOLUMEN_BY = -10f;
 	public static final String SPEAK = "Hello!";
 
 	//Constants Control
